cleaning up code and distributing across multiple files
